resole add and update item
